Split out bottom sheet component
